feat: add tests for camera presets including valid, invalid, and default behavior
